[API Compatibility] paddle.nn.functional.grid_sample sink into C++ (#76501)
* support grid_sample compatibility

* fix: update doc
