+++ paddle/fluid/pybind/arg_pre_process.h
@@ -44,6 +44,17 @@ void LogsumexpPreProcess(Value* x, std::vector<int>* axis, bool* reduce_all);
 
 void SumPreProcess(Value* x, Value* axis);
 void IsClosePreProcess(Value* x, Value* y, Value* rtol, Value* atol);
+
+void GridSamplePreProcess(Tensor* x,
+                          Tensor* grid,
+                          std::string* mode,
+                          std::string* padding_mode,
+                          bool* align_corners);
+void GridSamplePreProcess(Value* x,
+                          Value* grid,
+                          std::string* mode,
+                          std::string* padding_mode,
+                          bool* align_corners);
 }  // namespace pybind
 
 }  // namespace paddle
