+++ paddle/phi/ops/yaml/python_api_info.yaml
@@ -271,3 +271,10 @@
     use_default_mapping : True
   pre_process:
     static_func: IsClosePreProcess(x, y, rtol,atol)
+
+- op : grid_sample
+  name : [paddle.nn.functional.grid_sample]
+  args_alias:
+    x : [input]
+  pre_process:
+    func : GridSamplePreProcess(x, grid, mode, padding_mode, align_corners)
